Refactor: Changed the ExamRepository to use the new comparers
